video_agent_mvp/backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart==0.0.6
websockets
google-generativeai
numpy 